+++ tests/test_tracer.py
@@ -464,6 +464,34 @@ class DepsTest(AbstractTracerTest):
             },
         )
 
+    def test_geometry_nodes_node_input(self):
+        """Test linked collection as input to a geom node.
+
+        Here a Geometry Nodes node references a collection that is not instanced
+        into the scene, which caused it to be missed.
+        """
+        self.assert_deps(
+            "geometry-nodes-3/shot_file.blend",
+            {
+                b"LIset_file.blend": Expect(
+                    type="Library",
+                    full_field="name[1024]",
+                    dirname_field=None,
+                    basename_field=None,
+                    asset_path=b"//set_file.blend",
+                    is_sequence=False,
+                ),
+                b"LIlib_trash.blend": Expect(
+                    type="Library",
+                    full_field="name[1024]",
+                    dirname_field=None,
+                    basename_field=None,
+                    asset_path=b"//lib_trash.blend",
+                    is_sequence=False,
+                ),
+            },
+        )
+
     def test_usage_abspath(self):
         deps = [
             dep